requirements.txt
flask==3.0.3
python-dotenv==1.0.1
mailjet-rest==1.3.4
gspread==6.1.2
google-auth==2.33.0
itsdangerous==2.2.0
